# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ingestor System API
  description: |-
    RESTful API for the Ingestor System that provides access to content processing,
    entity extraction, and database capabilities using Claude AI.
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Local development server
...
paths:
...
  /content:
    get:
      tags:
        - content
      summary: List content items
      description: Returns a paginated list of content items with optional filtering
      operationId: listContent
      parameters:
        - name: type
          in: query
          description: Filter by content type
          schema:
            type: string
            enum: [text, image, video, code, pdf, document]
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: Field to sort by
          schema:
            type: string
            default: createdAt
            enum: [createdAt, updatedAt, title, type]
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            default: desc
            enum: [asc, desc]
        - name: q
          in: query
          description: Search query
          schema:
            type: string
      responses:
        '200':
          description: List of content items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    post:
      tags:
        - content
      summary: Create content
      description: Creates a new content item
      operationId: createContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentCreate'
      responses:
        '202':
          description: Content created and processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentCreatedResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /content/upload:
    post:
      tags:
        - content
      summary: Upload content file
      description: Uploads a file as content
      operationId: uploadContent
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: File uploaded and processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentCreatedResponse'
        '400':
          description: Invalid input or file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /content/{id}:
    get:
      tags:
        - content
      summary: Get content
      description: Returns a specific content item
      operationId: getContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Content item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    put:
      tags:
        - content
      summary: Update content
      description: Updates a content item's metadata
      operationId: updateContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - metadata
              properties:
                metadata:
                  type: object
                  properties:
                    title:
                      type: string
                    description:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
      responses:
        '200':
          description: Content updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    delete:
      tags:
        - content
      summary: Delete content
      description: Deletes a content item
      operationId: deleteContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Content deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Content with ID 123e4567-e89b-12d3-a456-426614174000 deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /content/{id}/entities:
    get:
      tags:
        - content
        - entities
      summary: Get content entities
      description: Returns entities associated with a content item
      operationId: getContentEntities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of entities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /content/{id}/related:
    get:
      tags:
        - content
      summary: Get related content
      description: Returns content items related to the specified content
      operationId: getRelatedContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of related content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  
  #
  # Entity Management Endpoints
  #
...
  /entities/{id}/content:
    get:
      tags:
        - entities
      summary: Get entity content
      description: Returns content items associated with the specified entity
      operationId: getEntityContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of content items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
...
components:
  schemas:
...
    ContentCreate:
      type: object
      required:
        - content
        - type
      properties:
        content:
          type: string
          example: This is a sample text document about artificial intelligence.
        type:
          type: string
          enum: [text, image, video, code, pdf, document]
          example: text
        filename:
          type: string
          example: sample.txt
        metadata:
          type: object
          properties:
            title:
              type: string
              example: Sample Text Document
            description:
              type: string
              example: A sample document for testing
            tags:
              type: array
              items:
                type: string
              example: [sample, ai]
        processingOptions:
          type: object
          properties:
            extractEntities:
              type: boolean
              default: true
            enableChunking:
              type: boolean
              default: true
            chunkSize:
              type: integer
              default: 500000
              minimum: 1000
              maximum: 1000000
            chunkOverlap:
              type: integer
              default: 5000
              minimum: 0
              maximum: 50000
            chunkStrategy:
              type: string
              enum: [size, paragraph, sentence]
              default: size
...
    ContentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentResponse/properties/data'
        meta:
          type: object
          properties:
            pagination:
              type: object
              properties:
                limit:
                  type: integer
                  example: 20
                offset:
                  type: integer
                  example: 0
                total:
                  type: integer
                  example: 156
                next:
                  type: string
                  example: /api/v1/content?limit=20&offset=20
                  nullable: true
                previous:
                  type: string
                  example: null
                  nullable: true
...
    EntityListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation error
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                    example: content
                  message:
                    type: string
                    example: Content is required
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key